Shadow Configuration Change
Changed the ShadowConfiguration to made it in the future easier to write and access.

Now, not everywhere you will "feel" the change.
