+++ Resonans/Views/OnboardingView.swift
@@ -106,7 +106,7 @@ struct OnboardingFlowView: View {
                         RoundedRectangle(cornerRadius: 36, style: .continuous)
                             .stroke(primary.opacity(AppStyle.strokeOpacity), lineWidth: 1)
                     )
-                    .shadow(DefaultShadowConfiguration.large.configuration(for: colorScheme))
+                    .shadow(ShadowConfiguration.largeConfiguration(for: colorScheme))
 
                 VStack(spacing: 18) {
                     Image(systemName: "waveform.circle.fill")
@@ -228,7 +228,7 @@ struct OnboardingFlowView: View {
                 RoundedRectangle(cornerRadius: 28, style: .continuous)
                     .stroke(primary.opacity(AppStyle.strokeOpacity), lineWidth: 1)
             )
-            .shadow(DefaultShadowConfiguration.medium.configuration(for: colorScheme))
+            .shadow(ShadowConfiguration.mediumConfiguration(for: colorScheme))
             Spacer()
         }
     }
@@ -346,7 +346,7 @@ private struct FavoriteSelectionCard: View {
                             .stroke(primary.opacity(isSelected ? 0.35 : AppStyle.strokeOpacity), lineWidth: isSelected ? 2 : 1)
                     )
             )
-            .shadow(DefaultShadowConfiguration.medium.configuration(for: colorScheme))
+            .shadow(ShadowConfiguration.mediumConfiguration(for: colorScheme))
         }
         .buttonStyle(.plain)
     }
